update-geoip: allow reporting current status and not updating

+++ update-geoip.sh
@@ -1,13 +1,38 @@
 #!/bin/sh
 set -x
 set -e
+
+update=false
+status=false
+while [ $# -gt 0 ]; do
+	case "$1" in
+	update|-update|--update)
+		update=true
+		shift
+		;;
+	status|-status|--status)
+		status=true
+		shift
+		;;
+	*)
+		break
+		;;
+	esac
+done
+
 pkgs='GeoIP-db-City GeoIP-db-Country GeoIP-db-IPASNum xtables-geoip'
 for pkg in ${*:-$pkgs}; do
+	$status && continue
+
 	./builder -g -ns $pkg
 	cd $pkg
-	rm -vf *.gz *.zip
+
+	$update && rm -vf *.gz *.zip
+
 	specfile=*.spec
+
 	../md5 -p1 $specfile
+
 	version=$(awk '/^Version:[ 	]+/{print $NF}' $specfile)
 	if [ $pkg = "xtables-geoip" ]; then
 		dt4=$(TZ=GMT stat -c '%y' *.zip | awk '{print $1}' | tr -d -)
@@ -29,5 +54,11 @@ for pkg in ${*:-$pkgs}; do
 	fi
 
 	../builder -bb *.spec
+	cd ..
+done
+
+for pkg in ${*:-$pkgs}; do
+	cd $pkg
+	git status -s
 	cd ..
 done
